Moved read_commandline into janitor.io   deleted old read_commandline.py file
